LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/style-inheritance-expected.txt


FAIL style inheritance of pseudo elements assert_equals: wrapper expected "linear" but got "ease"
